requirements.txt
qinfo-python==0.1.0
nuitka